+++ packages/combobox/src/Combobox.story.tsx
@@ -14,7 +14,6 @@ import {
 
 const wrapperStyle = css`
   width: 256px;
-  height: 100vh;
   padding-block: 64px;
 `;